==> 1_Notebooks/1_xETL.ipynb <==
# Conexión entre Google Colab y Drive
from google.colab import drive
drive.mount('/content/drive')

# **<font color="#ff5653">Extract, transform and load</font>**

Archivos de Steam:
- steam_games.json.gz
- users_items.json.gz
- user_reviews.json.gz

Se llevara a cabo la apertura de los archivos, EDA preliminar y ETL hasta obtener un dataframe de cada archivo limpio para ser procesado posteriormente.

# Libreria
import pandas as pd
import gzip                            # Archivos_comprimidos.gz
import json                            # Archivos.json
from textblob import TextBlob          # Analisis de sentimiento
import csv                             # Archivos.csv
import ast                             # Gramatica de sintaxis abstracta - Abstract Syntax

## **<font color="#ff5653">steam_games.jso.gz</font>**

# Funcion para descomprimir, abrir y generar el DataFrame
def steam_games(archivo_json_gz):
    data = []
    with gzip.open(archivo_json_gz, 'rb') as f:
        for line in f:
            line_content = line.decode('utf-8')
            data.append(json.loads(line_content))
    df_steam_games = pd.json_normalize(data)
    return df_steam_games

# Llamamos a la función
df_steam_games = steam_games("/content/drive/MyDrive/Datasets_Steam/steam_games.json.gz")

Exploratory Data Analysis (EDA) - Análisis Exploratorio de Datos

# Vemos el DataFrame
df_steam_games

# Informacion de no nulos y tipos de datos.
df_steam_games.info()

# Suma de valores nulos (NaN) en cada columna
df_steam_games.isna().sum()

# Cantidad de filas y columnas
df_steam_games.shape

Extract, Transform and Load (ETL) - Extracción, Transformación y Carga.

# Copiamos el archivo original
df_steam_games1 = df_steam_games.copy()

# Eliminamos filas que solo contengan todos sus valores nulos.
df_steam_games1 = df_steam_games1.dropna(how='all')

# Vemos el DataFrame
df_steam_games1

# Cantidad de filas y columnas
df_steam_games1.shape

# Ordenamos el indice del DataFrame
df_steam_games1.reset_index(drop=True, inplace=True)

# Vemos el DataFrame con el indice ordenado
df_steam_games

# Guardamos el DataFrame en un archivo .csv

df_steam_games1.to_csv("steam_games_ETL.csv", index=False)

- Teniamos originalmente: 120.445 filas y 13 columnas.
- Ahora tenemos: 32.135 filas y 13 columnas.

## **<font color="#ff5653">users_items.jso.gz</font>**

# Funcion para descomprimir, abrir y generar el DataFrame
def users_items(archivo_json_gz):
    with gzip.open(archivo_json_gz, "rt", encoding="iso-8859-1", errors="ignore") as file:
        data = [ast.literal_eval(line) for line in file]
        df_users_items = pd.DataFrame(data)
    return df_users_items

# Llamamos a la función
df_users_items = users_items("/content/drive/MyDrive/Datasets_Steam/users_items.json.gz")

Exploratory Data Analysis (EDA) - Análisis Exploratorio de Datos.

# Vemos el DataFrame
df_users_items

# Informacion de no nulos y tipos de datos.
df_users_items.info()

# Suma de valores nulos (NaN) en cada columna
df_users_items.isna().sum()

# Cantidad de filas y columnas
df_users_items.shape

Extract, Transform and Load (ETL) - Extracción, Transformación y Carga.

# Copiamos el archivo original
df_users_items1 = df_users_items.copy()

# Desanidamos la columna "items"
# Fragmentamos el archivo, para poder desanidarlo.

tamaño =1000
fragmentos = [df_users_items1[i:i + tamaño] for i in range(0, len(df_users_items1), tamaño)]

# Funcion para desanidar

def desanidar_items(df):

    items = df["items"].explode()
    df = df.drop("items", axis=1)
    df = pd.concat([df, items.apply(pd.Series)], axis=1)

    return df

# Llamamos a la funcion
fragmentos_desanidados = [desanidar_items(fragmento) for fragmento in fragmentos]

# Concatenar los fragmentos en un DataFrame
df_items_desanidado = pd.concat(fragmentos_desanidados)

# Vemos el DataFrame desanidado
df_items_desanidado

# Ordenamos el indice del DataFrame
df_items_desanidado.reset_index(drop=True, inplace=True)

# Guardamos el DataFrame en un archivo .csv

df_items_desanidado.to_csv("users_items_ETL.csv", index=False)

- Teniamos originalmente: 88.310 filas y 5 columnas.
- Ahora tenemos: 5.170.015 filas y 9 columnas.

## **<font color="#ff5653">user_reviews.json.gz</font>**

# Funcion para descomprimir y abrir
def user_reviews(archivo_json_gz):
    with gzip.open(archivo_json_gz, 'rt',encoding="utf-8") as file:
        return[ast.literal_eval(line.strip()) for line in file]

# Llamamos a la funcion
df= user_reviews("/content/drive/MyDrive/Datasets_Steam/user_reviews.json.gz")

# Creamos el DataFrame
df_user_reviews=pd.DataFrame(df)

Exploratory Data Analysis (EDA) - Análisis Exploratorio de Datos

# Vemos el DataFrame
df_user_reviews

# Informacion de no nulos y tipos de datos.
df_user_reviews.info()

# Suma de valores nulos (NaN) en cada columna
df_user_reviews.isna().sum()

# Cantidad de filas y columnas
df_user_reviews.shape

Extract, Transform and Load (ETL) - Extracción, Transformación y Carga.

# Copiamos el dataframe
df_user_reviews1 = df_user_reviews.copy()

# Cada elemento de la lista en la columna "reviews" se convertirá en una fila separada
df_user_reviews_explode = df_user_reviews1.explode("reviews")

df_user_reviews_explode

# Normalizacion
df_user_reviews_normalize = pd.json_normalize(df_user_reviews_explode['reviews'])

df_user_reviews_normalize

# Eliminamos los indices
df_user_reviews_explode.reset_index(drop=True, inplace=True)
df_user_reviews_normalize.reset_index(drop=True, inplace=True)

# Concatenamos los dos DataFrame
df_user_reviews_nuevo = pd.concat([df_user_reviews_explode.drop('reviews',axis=1),df_user_reviews_normalize], axis=1)

df_user_reviews_nuevo

# Informacion de no nulos y tipos de datos.
df_user_reviews_nuevo.info()

# Suma de valores nulos (NaN) en cada columna
df_user_reviews_nuevo.isna().sum()

# Cantidad de filas y columnas
df_user_reviews_nuevo.shape

- Teniamos originalmente: 25.799 filas y 3 columnas.
- Ahora tenemos: 59.333 filas y 9 columnas.

# Eliminamos nulos y rellenar con 1: si solo eliminamos los nulos nos quedaremos sin datos para realizar el analisis de sentimiento.
df_review = df_user_reviews_nuevo.fillna(1)

df_review.shape

df_review.isna().sum()

## **<font color="#ff5653">Sentiment_analysis</font>**

Analisis de sentimiento con NLP

En el dataset user_reviews se incluyen reseñas de juegos hechos por distintos usuarios.

Vamos a crear una columna "sentiment_analysis" aplicando análisis de sentimiento con NLP -Natural Language Processing- con la siguiente escala: debe tomar el **valor "0" si es malo**, **"1" si es neutral** y **"2" si es positivo**.

Esta nueva columna debe reemplazar la columna "review" para facilitar el trabajo de los modelos de machine learning y el análisis de datos. De no ser posible este análisis por estar ausente la reseña escri

# Instalamos la biblioteca: VADER (Valence Aware Dictionary and sEntiment Reasoner).
!pip install vaderSentiment

# Importamos la biblioteca
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

sentimiento = SentimentIntensityAnalyzer()

# Puntuamos el texto
puntuacion_sentimiento = []

for texto in df_review['review']:
    if isinstance(texto, str):
        s = sentimiento.polarity_scores(texto)
        puntuacion_sentimiento.append(s['compound'])
    else:
        puntuacion_sentimiento.append(0.0)

puntuacion_sentimiento[:10]

df_review_x = df_review

# Agregamos al dataframe una columna "puntuacion_sentimiento"
df_review_x["puntuacion_sentimiento"] = puntuacion_sentimiento

df_review_x.head(2)

Usamos la columna "puntuacion_sentimiento" para hacer la clasificacion.

- sentimiento positivo: puntuacion >= 0.05
- sentimiento netro: puntuacion entre -0.05 a 0.05
- sentimiento negativo: puntuacion <= -0.05

# Analisis de sentimiento
analisis_sentimiento = []

for puntuacion in df_review_x["puntuacion_sentimiento"]:
    if puntuacion >= 0.05:
        s = 2
    elif puntuacion > -0.05 and puntuacion < 0.05:
        s = 1
    else:
        s = 0
    analisis_sentimiento.append(s)

# Agregamos al dataframe una columna "analisis_sentimiento"
df_review_x["analisis_sentimiento"] = analisis_sentimiento

df_review_x.head(2)

# Eliminamos columnas que no vamos a usar
columnas_eliminar = ["review","puntuacion_sentimiento"]

df_user_reviews_limpio = df_review_x.drop(columns=columnas_eliminar)

df_user_reviews_limpio.head(2)

df_user_reviews_limpio.shape

# Guardamos todos los cambios en un archivo .csv
df_user_reviews_limpio.to_csv("user_reviews_ETL.csv", index=False)
