==> 1_Notebooks/2.2_xFunciones.ipynb <==
# Coneccion entre Google Colab y el Drive
from google.colab import drive
drive.mount('/content/drive')

# **<font color="#ff5653">Funciones</font>**

En este notebook se creara el archivo donde las funciones tomaran la información.

Archivos utilizados:
 - steam_games_ETL.csv
 - users_items_ETL.csv
 - users_reviews_ETL.csv

Funciones que tomaran la información:
- Users_Recommend
- Users_Not_Recommend
- Sentiment_analysis

Archivos creados:
- df_f34y5.csv
- df_f34y5.parquet

# Libreria
import pandas as pd
import numpy as np
import re                     # Expresiones regulares

### 1- Extraemos el año

Del df_user_reviews extraemos el año de la columna "posted" y lo colocamos en una nueva columna.

# Abrimos el archivo user_reviews_sentimen_analysis.csv
df_user_reviews = pd.read_csv("//content/drive/MyDrive/Datasets_Steam/user_reviews_ETL.csv")

# Columna "posted", vemos si hay valores nulos
df_user_reviews["posted"].isna().sum()

df_user_reviews.shape

# Utiliza una expresión regular para buscar años en formato "YYYY" en cada elemento de la columna "posted".

lista_posted = df_user_reviews["posted"]
lista_almacenaje = []
lista_años = r'\b\d{4}\b'       # Busca año en formato "YYYY"

for fechas in lista_posted:
    # Si el elemento es una cadena, buscar un año en formato "YYYY"
    if isinstance(fechas, str):
        yyyy = re.search(lista_años, fechas)
        if yyyy:
            lista_almacenaje.append(yyyy.group())
        else:
            lista_almacenaje.append('NaN')
    # Si el elemento no es una cadena, agregar 'NaN' a la lista de almacenaje
    else:
        lista_almacenaje.append('NaN')

# Vemos lo que extrajimos
lista_almacenaje[:10]

Para poder cocatenar la lista "lista_almacenaje" al dataframe df_user_review, debo tener la misma cantidad de filas.

De no tener las misma cantidad de filas los valores faltantes se rellenan con valores nulos.

En la "lista_almacenaje" aparecen "NaN", esto se debe que en la lista no poseen el año de publicado (ver ejemplos).

Y como el ultimo año de publicado es 2015, suponemos que estos datos fueron extraidos en 2016 por lo que donde hay "NaN" en "lista_almacenaje" voy a colocar 2016.

De esta forma al cocatenar "lista_almacenaje" a df_user_review, tendran las mismas cantidad de filas y evitamos que se agregen valores nulos automaticamente al dataframe.

Ejemplos:

- Sin año:
- indice=6 Posted February 3.
- indice=56 Posted August 23.

- Con año:
- indice=7 Posted December 4, 2015.
- indice=8 Posted November 3, 2014.
- indice=12 Posted October 14, 2013.
- indice=13 Posted July 28, 2012.
- indice=1 Posted July 15, 2011.

# Reemplazamos "NaN" por "2016"
lista_almacenaje_x = ["2016" if x == "NaN" else x for x in lista_almacenaje]

np.shape(lista_almacenaje_x)

np.shape(df_user_reviews)

# Concatenamos la lista al dataframe
df_user_reviews_x = pd.concat([df_user_reviews, pd.DataFrame({"año_publicado": lista_almacenaje_x})], axis=1)

df_user_reviews_x.head(2)

### 2- Extraemos el año

Del df_steam_games extraemos el año de la columna "release_date" y lo colocamos en una nueva columna.

# Abrimos el archivo steam_games_limpio.csv
df_steam_games = pd.read_csv("/content/drive/MyDrive/Datasets_Steam/steam_games_ETL.csv")

# Suma de valores nulos de la columna "release_date"
df_steam_games['release_date'].isna().sum()

# Funcion para extraer el año
def año_lanzamiento(fecha):
    if pd.notna(fecha):
        años = re.search(r'\b\d{4}\b', fecha)
        if años:
            return años.group()
    return np.NaN

# Llamamos a la funcion: solo para la columna "release_date"
año_lanzamiento = df_steam_games["release_date"].apply(año_lanzamiento)

año_lanzamiento.head(5)

# Concatenamos la columna "año_lanzamiento" con el DataFrame df_steam_games
df_steam_games_año = pd.concat([df_steam_games, pd.DataFrame({"año_lanzamiento": año_lanzamiento})], axis=1)

df_steam_games_año.head(2)

### 3- Horas de juego por genero

De df_users_items sumamos las horas de juego de "Playtime_forever" agrupados por los ID de la columna "item_id" tendremos las horas de juego para cada genero.

# Abrimos el archivo users_items
df_users_items = pd.read_csv("/content/drive/MyDrive/Datasets_Steam/users_items_ETL.csv")

# Agrupamos los ID similar de las columnas "item_id" y "user_id", para despues se suman las horas de la columna "playtime_forever" de cada columna agrupada
columnas_df_users_items = df_users_items.groupby(["item_id", "user_id"])["playtime_forever"].sum().reset_index()

columnas_df_users_items.head(2)

# Unimos los dataframe columnas_df_users_items con df_user_reviews_x
df_items_y_reviews = pd.merge(columnas_df_users_items, df_user_reviews_x, on=["item_id", "user_id"], how="inner")

# Tenemos en un dataframe "item_id", la suma de "playtime_forever", agrupados por "user_id", el año en que el usuario jugo "año_publicado" y que hizo la reseña "analisis_sentimiento"
df_items_y_reviews.head(2)

# Cambiamos en df_steam_games_año el nombre de la columna a "id" a "item_id" para poder unir a otro dataframe por "item_id"
df_steam_games_año.rename(columns={"id": "item_id"}, inplace=True)

# Unimos los dataframe df_items_y_reviews con df_steam_games_año
df_xxx = pd.merge(df_items_y_reviews, df_steam_games_año, on="item_id", how='inner')

df_xxx.head(2)

df_xxx.info()

### 4- Columnas requeridas para las funciones:

- item_id
- title
- año_lanzamiento
- genres
- user_id
- playtime_forever
- recomend
- analisis_sentimiento
- año_publicado

# Creamos columnas a extraer
columnas_x9 = ["item_id","title","año_lanzamiento","genres","user_id","playtime_forever","recommend","analisis_sentimiento","año_publicado"]

# Creamos nuevo dataframe con las columnas seleccionadas
df_f34y5 = df_xxx[columnas_x9]

df_f34y5.head(2)

df_f34y5.info()

# Cambiamos el tipo de datos de la columna "title" de object a string
df_f34y5["title"] = df_f34y5["title"].astype("string")

# Cambiamos el tipo de datos de la columna "año_publicado" de object a int64
df_f34y5["año_publicado"] = df_f34y5["año_publicado"].astype("int64")

# Eliminamos valores nulos de "año_lanzamiento" para poder cambiar el tipo de dato.
df_f34y5.dropna(subset="año_lanzamiento",inplace=True)

# Cambiamos el tipo de datos de la columna "año_lanzamiento" de object a int64
df_f34y5["año_lanzamiento"] = df_f34y5["año_lanzamiento"].astype("int64")

df_f34y5.info()

df_f34y5.shape

# Guardamos el dataframe obtenido en formato .csv y parquet
df_f34y5.to_csv("df_f3f4yf5.csv", index=False)

df_f34y5.to_parquet("df_f3f4yf5.parquet", index=False)
